Refactor Directories Employee
