+++ backend/src/main/java/de/neuefische/cgnjava222/zoomanagement/zoo/employee/Employee.java
@@ -1,4 +1,4 @@
-package de.neuefische.cgnjava222.zoomanagement.zoo.employee.model;
+package de.neuefische.cgnjava222.zoomanagement.zoo.employee;
 
 import org.springframework.data.annotation.Id;
