+++ frontend/src/animal/Animal.ts
