+++ frontend/src/animal/AnimalList.tsx
