+++ frontend/src/employees/AddEmployee.tsx
@@ -1,45 +1,32 @@
 import {ChangeEvent, FormEvent, useState} from "react";
-import {Employee} from "../model/Employee";
+import {Employee} from "./Employee";
 import {showError} from "./ErrorEmployee";
-
 
 type AddEmployeeProps = {
     addEmployee: (name: string) => Promise<Employee>
 }
 
-
 export default function AddEmployee(props: AddEmployeeProps) {
-
-
     const [name, setName] = useState<string>("");
-
     function onNameChange(event: ChangeEvent<HTMLInputElement>) {
         setName(event.target.value)
     }
 
     const onEmployeeSubmit = (event: FormEvent<HTMLFormElement>) => {
         event.preventDefault();
-
         if (!name) {
             console.error("Name must not be empty")
-
         } else {
-
-            props.addEmployee(name).then(s => setName("")).catch(error => showError());
-
+            props.addEmployee(name).then(() => setName("")).catch((error) => showError(error));
         }
     }
 
     return (
         <div>
-
             <form className="form" onSubmit={onEmployeeSubmit}>
-
                 <input value={name} onChange={onNameChange}/>
                 <button>Hinzufügen</button>
             </form>
-
-
         </div>
     )
 }
