+++ frontend/src/trucks/Truck.ts
