+++ frontend/src/trucks/TruckGallery.tsx
